requirements.txt
django==1.8.2
django-debug-toolbar
django-admin-tools
django-celery
redis
celery[redis]
pillow
openpyxl